backend/pyproject.toml
[tool.poetry]
name = "aqua"
version = "0.1.0"
description = "Application for tracking your water balance"
authors = ["emptybutton <88573504+emptybutton@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
typenv = "^0.2.0"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
ruff = "^0.4.5"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"

[tool.poetry.group.fix.dependencies]
certifi = "^2024.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
no_namespace_packages = true

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.lint]
select = [
    "ANN", "F", "W", "E", "B", "C90", "N", "UP", "YTT", "ASYNC", "PL", "RUF",
    "Q", "RET", "SLF", "SLOT", "SIM", "TID252", "TCH", "ARG", "PTH", "ERA", "TRY",
    "PERF", "INP"
]
ignore = ["ANN101", "ANN102", "N818"]
